+++ src/auth/mod.rs
@@ -1,0 +1,4 @@
+pub mod git_auth;
+mod git_http;
+mod git_ssh;
+mod ssh_utils;
