+++ src/events/git_clone.rs
@@ -1,7 +1,7 @@
 use super::AtomicEvent;
+use crate::auth::git_auth::setup_auth_callbacks;
 use crate::bgit_error::{BGitError, BGitErrorWorkflowType, NO_RULE, NO_STEP};
 use crate::rules::Rule;
-use crate::utils::git_auth::setup_auth_callbacks;
 use std::env;
 use std::path::Path;
