+++ src/events/git_push.rs
@@ -1,7 +1,7 @@
 use super::AtomicEvent;
+use crate::auth::git_auth::setup_auth_callbacks;
 use crate::bgit_error::{BGitError, BGitErrorWorkflowType, NO_RULE, NO_STEP};
 use crate::rules::Rule;
-use crate::utils::git_auth::setup_auth_callbacks;
 use git2::Repository;
 use log::debug;
 use std::path::Path;
